Move subclasses to impl packages, finalize Framework
I decided to leave implementations of all the framework stuff, as it can not only help the end-user find ways to create their own implementations, but it also helps in making it easier to use the framework. Regardless, using impl classes are not the recommended way to go, and it's preferred that you implement your own.