Update module, fix ItemRenderer, new PacketEvent, Blink added
Currently, blink will not work. Effectively, when Blink is disabled, it attempts to send all the packets through NetworkManager#sendPacket, however, the issue is that we cannot guarantee that it has actually been sent, or if it's being held up in the outbound queue (which means it'll be sent later). As a result, we might move the player forward, but the packets themselves will be sending data for the past, which glitches the player position. We need to find a way to either force send packets, or make up for the difference between the last sent and the current position. We need to find someway to avoid being delayed, effectively.