Allow deprecated reflection if debugging, move debug to client
Blink is still broken, but not for the reason I thought. Although the packets will now be forced into an order, they will still cause problems on public servers. This requires further testing, as well as finding out how to bypass server protections (it may have something to do with the speed at which packets are sent)