Update Timer
Irrelevant to this commit, but after testing, I'm certain servers are blocking blink because the packets are sent too fast. Either they can't process them fast enough, or they flat out reject them, but regardless, it's breaking Blink. The issue is that we cannot delay without harming user experience, so we will need to think of a better way to handle it.