Fix Mixin not correctly implementing, fix Ghost static instance
Mixin wasn't enforcing intrinsic simply because I wasn't @Implement-ing it. Simple mistake, yet it led to devastating consequences. Why did it allow it to work in development is beyond me.

As for ghost, due to the nature of Gson#fromJson, I ended up replacing Ghost.self with the reference of the deserialized object instead of the one I intended. This led to problems, so in order to avoid that, I added a null check.