Implement 'Release v2.1-alpha' #6
* Change ForgeMod logger to prefix with ${name}

This makes more sense, given that the rest of the log entries refer to our client via it's name, not the modid.

* Change config to use JsonParser

Gson just isn't really appropriate here, and although it works, it makes more sense, given the context of what we are doing, to use a JsonParser

* Privatize reflectionSearch in manager

No one can call it, since ImmutableSet is already defined as itself by construction, making it pretty useless

* Update ModInfo javadocs

Specifies a bit more into what goes on inside of ModInfo.

* Add tracing to Logger

This can greatly help when debugging.

* Change MixinNetworkManager to redirect call

This is a lot more logical, and less intrusive

* Change MixinGameSettings to Inject

I.. don't even know why this was an Overwrite from the start.

* Force runClient to use 1G of ram

This is moreso for testing purposes, as I want to focus on memory issues within the client.

* Implement some prplz fixes

This may not actually be allowed by prplz. Let's treat this commit with some caution.

* Update mcmod.info loading

This exists entirely due to the way Fabric loads mods. Upon testing, Fabric actually loads it's default fabric.mod.json, rather than the one we specify, which results in missing properties. In order to circumvent that, we must directly manipulate the jar file.

* Add Fabric support to ModInfo

Thankfully it's not a big change

* Prevent console while Ghost is active

This is something I should've already handled. Oh well.

* Remove dependency on Minecraft directory paths

We don't need it from minecraft, as the jar directory is already set to that location. As a result, we don't need to access it directly, but can infer it using File

* Fix resource packs failing to load

What an odd bug. I must investigate, since this feels like such an unecessary Overwrite

* Update NameProtect to obfuscate instead of jumble

This looks much better imo.

* Update KeyBindScreen to only disable explicitly

This can help a user revert if they accidentally pressed the bind screen.

* Update creation of ModInfo to support Fabric

Fabric can now properly load this framework very easily. It's basically the same thing, except you must use some ModInitializer or ClientModInitializer instead of a @EventHandler thing.

* Update Configurable

Still don't recommend this class, however it should work, given the implementation, if you really wanted

* Trace method calls only when debug in logger

It's very useless for anything other than debugging.

* Finalize some classes

These classes aren't meant to be extended, so yeah.

* Move InfinityScreen

It's not dependant on any module stuff, so it can be moved out.

* Update ChatUtils to fix appending colors in NameProtect

Colors were defaulting since a reset was leaking. Since it's annoying, it's been fixed. God, I hate the ChatComponent stuff

* Deprecate NameProtect

ChatComponent has too much stupid behaviour, I'm going to rewrite it

* Reformat code

* Update ChatUtils, fix color formats resetting via protect

Problematically, ChatComponentText appends a reset to it's formatted text, which obviously causes problems with the coloring of the text. This fixes that

* Fix nameprotect overriding original colors

ChatUtils shouldn't assume the default style is white, even if a null color translates to white, it doesn't mean that it is white. This just inhibits the parent styles from being able to set their colors

* Comply with CodeFactor

* Update JavaDocs

* Finalize release