Implement 'Release v2.2-alpha' #7
* Move timer into separate package

Cleaning up the utilities

* Fix AbstractResourcePack incorrectly loading icons

I was referencing the image that was unscaled, and not drawing to the scaled one, causing obvious issues

* Force Ghost when OBS is running

We need to test to see if this would work on windows. Additionally, we need to give the user a way to re-enable ghost if they want, we need to warn the user in a much more logical manner (not everyone has logs on a second monitor), we need to fix some handling with active states, and we need to improve detection for more recorders

* Reserve OS data as a static value

There is 0 reason to retrieve it each time, that is just tedious

* Move environment information to it's own class

This can help us create environmental changes that are necessary for specific modding api.

* Error handle for when info json cannot be found

This isn't the best way to handle it, but it will work for now.

* Modify how the modding environment works

This is a lot more logical, and will help once supporting larger and more complex formats

* Fix info file not loading into the client

I had assumed refactor targeted by build script, but it looks like that wasn't the case

* Change ModEnvironment to make use of bytes instead

This makes less use of an OOP pattern, and a more, C-like approach. This should save on unnecessary memory usage with environment objects.

* Add support for opening url

Basic stuff

* Rewrite configurable to sport less nonsense methods

Lots of them aren't needed. Frankly, I'm considering allowing this to be used, but I still don't recommend it, since reflection is slow.

* Finalize certain classes

They shouldn't be overridden

* Fix MixinEntityPlayer not normalizing rotations

This is essential to replicate the fix that prplz commited (he directly referenced Entity#getLook)

* Fix performance issues with Ghost

All of it is now managed by it's own thread

* Implement AutoTool

This must be modified to support enchanted items

* Remove Wrapper call to mouseOver, remove dependency, fix History

History will now display names in an order, additionally, we no longer need the whole mojang API + all it's annoying dependencies, since we can do exactly all that with normal java stuff.

* Add automatic refreshing to the gui multiplayer

Also add some optimizations to the refreshing

* Create an infinite thread, and move obsChecker to static threads

Self-explanatory

* Implement execute command

Self explanatory, really, although it's a bit weird to be sure.

* Fix for codefactor stuff

Release has been finalized