Implement 'Release v2.3-alpha' #8
* Import Lombok into the gradle script

This isn't meant to be compiled into the code.

* Initial Lombok changes

The code is already looking much better. These are some things that need cleanup, alongside needing to figure out how val/var can be enabled, or whatever

* Update lombok dependency, move things for the new plugins dsl

I dont really like it per-se. I wonder if I've done it wrong..

* Move InfiniteThread, remove @NonNull from Drawables

We don't actually need it at all

* Add more sneaky throws

Cool stuff.

* Remove final class specification

Lombok does that automatically

* Modify manager to use a ImmutableMap

This helps specifically with Manager#get

* Add logger entry

We want to warn the dev if registrations occur with the same type twice, since it means the previously registered one is gone.

* Update travis, add System Architecture

Travis probably won't work. Thats fine

* Update travis to Gravis-CI

Yep

* Revert travis

Back to square 1.2

* Revert travis

...

* Add var/val usages

This wasn't working before, but it suddenly started working. I think it probably had something to do with cb1c78435f36690cc22e3bf2eb2ae1a3c872fd17

* Fix replacements['mcversion'] being null

This is due to the fact the property was not initialized when we define our replacements

* Fix gradle not correctly replacing

This should not cause problems now.

* Add FieldDefault

I should probably stop on using too much Lombok. If I ever decide to stop using it, it could cause so many issues

* Remove Minecraft Wrapper

It's becoming more of a chore to work with.

* Proof-of Concept

Yeah,so although the IDE will complain, this is completely valid in the context of Lombok.

* Revert "Proof-of Concept"

It was just a proof of concept

* Add Unsafe into the Agent

Because why not, frankly?

* Add Manifold to the project

I will likely NOT use the full manifold project, however I will use some parts, especially the preprocessor

* Make use of manifold stuff

Now we don't need to annotate with @SneakyThrows everywhere, as Manifold cleans that up for us. Additionally, we can now use String templates, which is a fantastic change.

* Deprecate Access Transformers

We can make use of @Jailbreak instead

* Use jailbreak extension

That is definitely a good addition.

* Remove String.formats and make use of extensions

* Add more @Jailbreak, remove catches

Some catches have been left behind, because these are actually necessary for certain functionality.

* Bring back access transformers

I seem to have forgotten that Minecraft has mappings...

* Fix AutoTool not working

This needs to be resolved ASAP

* Update travis to test the actual minecraft client

Yeah, basically the comment

* Remove runClient from gradle

Silly me, travis doesn't have a display configured. I don't even think they have a display manager actually.

* Import Google checkstyle

Tis a test, nothing more.

* Try to prevent forge code calls

And hopefully code factor stops complaining about unsafe..

* Revert

Hmm

* Create an OpenGL State machine wrapper

This will be expanded upon in the next release.

* Optimize Imports, remove Unsafe reference

Once again, MinecraftAgent will be worked on soon, as well as Unsafe stuff.

* Finalize, fix lack of import

IntelliJ optimizing of imports fails to recognize that the manifold strings thing needs them, so it removes the essential imports.

* Change Minecraft replacement to prevent manifold interfearance

Basically the title.