Create a memory allocation framework
This is the start of real performance improvements.

There are a couple of things we can improve here. First, create a much more broader sizeof method, which can parse classes (and only classes, this sizeof method does not calculate headers and padding, but rather the literal size of the object), as well as supporting slightly more optimized allocation systems (maybe remove MemoryObject in favour of a generic object?).