Implement 'Release v2.4-alpha' #9
* Update README to include manifold

Forgot to do this in the original release.

* Update OpenGLState wrapper

This *should* fix the issue with applying a blend func twice, causing textures to occasionally break.

* Replace usage of render events

In preparation of becoming more independent of forge.

* Make local variables to be inferred via lombok

* Add String extension, change code for debugging

Debugging can interfere with normal code, so we will split that up.

This also separates the use of runClient and genIntelliJRuns.

* More extension stuff

Yep, yep

* Rename extension method, fix stack overflow

Replace all is not the way to go here.

* Remove usage of genIntelliJRuns

* Modify OverlayEvent, upscale Infinity text, update version in debug

* Remove usage of OpenGLState 

We will improve it, then implement.

* Updating extension stuff

* Create a memory allocation framework

This is the start of real performance improvements.

There are a couple of things we can improve here. First, create a much more broader sizeof method, which can parse classes (and only classes, this sizeof method does not calculate headers and padding, but rather the literal size of the object), as well as supporting slightly more optimized allocation systems (maybe remove MemoryObject in favour of a generic object?).

* Remove javadocs from travis

They are absolutely incompatible with the manifold preprocessor.

* Fix ChatUtils switch case falling to default

There's probably more bugs like this hidden in the code.

* Move Mixins

This is for a specific purpose

* Rework MemoryPool

This will be worked on in the next release. This aims to make use of a universal memory pool to store and pull data from. It will greately simplify the nonsense code that we write.

* Enable Infinity Screen, prevent unecessary function calls

I failed to address drawables, although it was a priority for this release. I will definitely have to make that a priority.

* Remove OBS checking feature

Oh lawd, it's so terribly inefficient. I might bring it back if I have a better implementation for it.

* Improve ClickGui stuff

I honestly wish to rewrite all this, and abstract some common functionality into the infinity package (not the framework, unless I can make a font renderer).

* Revert MixinGuiMultiplayer refresh change

It's causing crashes, and failing to refresh. I'll address this much later, as it's such a low priority feature.

* Modify the implementation of Xray and Fullbright

This is less impractical

* Fix InfiniteThread crashes, remove AutoClicker needed by AutoTool

* Finalize release

Sorry MemoryPool, you will be moved to another project, as you are not really something I consider appropriate for a minecraft framework.

* Fix Fullbright

Yep, I had a feeling that would not go well.

* Fix TickingEntity something

Minecraft code...