Implement 'Release v3.0-beta' #10
* Move InfinityScreen package

* Remove useless isActive check from autotool

* Create TextRenderer, add some Client information as a module

I'd prefer if all text rendering was directed through text renderer at some point.

* Remove TimeUnit and update Info

I can't believe I never knew that there was a TimeUnit class already built into Java.

* Have TextRenderer handle if debugInfo

* Add glAttrib to breadcrumbs, fix Info from NPE

* Update javadocs

* Reformat, create WORKING_DIRECTORY variable

This gives us more direct influence over which directory we mess with.

* Add new commands

Dump will help me hopefully find more issues.

* Add TPS to Info

* Fix file spacing bug, and change up Timer formatting

* Add tons of debug mesasges, change framework logs to instance based

This can help specify exactly what "Manager" is calling the logs and such.

* Move Module Loggers to the parent

* Fix typo

I had originally moved the logging entry from onRender, and just copy pasted the x.getSimpleName()

* Fix incorrect timer calculations

* Test and update debugs for modules

They should provide some valuable feedback for future stuff

* String#replaceAll with String#replace

The naming for these methods is honestly stupid. Regardless, since we aren't making use of the regex support in replaceAll, we can safely switch to replace.

* Fix stream preventing reuse

Odd, but nothing that can't be fixed with slightly more verbosity.

* Create Random class

Is it necessary? Maybe.

* Changing some properties to static

We don't need to waste memory creating a new object with it's own fresh reference. We can very safely reuse references.

* Small change

...

* Update style.xml

This makes use of https://github.com/google/styleguide, modified for my preferences.

* Remove checkstyle

I'll add a more personal one later on.

* Remove OpenGLState

I have no actual reason to immediately implement it

* Update Lombok

When did we downgrade, or was it always at 1.18.8?

* Fix Fullbright issues with shaders

I believe the regression should be long gone.

* Update Manifold, use new stuff

Yay

* Change Overwrite to Redirect

* Force parameters to be immutable

Just a preference thing

* Revert usage of Ranges in Manifold

They are breaking Lombok stuff for now.

* Cleanup

* Update manifold, clean imports, move code around

* Change modding environment system

This is more open to changes, and can be implemented by others as well to support their own wacky environments

* Subtle improvements to code

Authors are handled differently in ModInfo, AutoTool won't spam you with messages if you're using the right tool, and Client uses fileOnExit

* Deprecate Drawable stuff in favour of Graphics

Drawable stuff will be completely removed soon, and replaced entirely with the graphics stuff. Alongside that, many more small things will be added to Graphics, such as TextRendering, Texture support, and more.

* Modify console to use new graphics

The code has also been cleaned up, kinda.

* Add ColorFactory, update Renderable's to hold their own color

This was mainly done for Button and modules, which will be fixed next. ColorFactory was created simply so that we don't have duplicate instances floating around in memory. The obvious issue is that it won't go away unless manually cleared.

* Remove Drawable, update ColorFactory, update ClickGui

There is a lot of code to clean up with the gui in this..

* Return gc call for freeMemory

Since freeMemory actually needs it.

* Change parameters to final, add precondition messages

* Create a MinecraftMod interface

The plan is to handle all loading through it.

* Remove the weird debug loading

It doesn't make sense to change up so much simply for debugging. The user should be able to understand what it means to debug, enabling it is not accidental.

* Implement EventExecutor, remove String.format stuff

It will remain in the logger, but it won't actively be used.

* Change Event to abstract

I mean, duh

* Change Event to be of a class type

This was actually a typo, that's being fixed right now.

* Update manifold

* Remove Event nonsense

Efkolia in the works will replace the in-built framework. It's not ready yet, but once it's done, it will be very useful

* Add health tags

* Style guide shift

Imports are now lazy, for minimalization

* Improve AutoClicker randomization

Since the cps range is ungodly low, it would generate the exact same 2-3 numbers in a random sequence, making it very easily detectable. This improves the randomization.

* Rename package

Also a style guide

* Fix package problems

Thanks IDEA, very nice refactor.