Rewrite Infinity from Forge -> Fabric
* Replace workspace with Themelio

This is the start of rewriting Infinity for Fabric

* Remove necessity on Fabric loading

We can load our own mod via mixins. We do not need the fabric system to do such. This gives us greater control over the lifecycle of the mod and its effects.

* Update some annotations

I'd like to keep these in as many places as possible. Creates a sense of null safety and makes the code coherent as to where a null is acceptable, versus where it isnt

* Change the way window titles work

Honestly, it doesn't really matter immediately, but it will matter once Infinity changes the name of the window. Since Infinity controls the window name, it needs it to remain constant in easily predictable ways. Having it change based on literally anything is quite stupid and unnecessary.

* Update Efkolia

The '+' versioning pings maven search, which takes an unfortunate amount of time to update. As such, I'll prefer to use specific versions.

* Remove nonNull causing an error

This is an outdated configuration and is no longer necessary.

* Change mixin for new style 

All mixins should be package-private and abstract. This represents that they are not meant to be accessed beyond the scope of their package, and that they are not meant to be instantiated.

* Update Efkolia

* Change library attachments

This helps us attach any other libraries very easily

* Resolve imports

* Update Efkolia

God the procrastination

* Create a basic Module framework for Infinity

It finally works. I can rest easy

* Target all getWindowTitle references

I don't know why Mojang is doing what its doing, but no.

* Add key event, Ghost module, save/load config, and refactor

Oh yeah, its all coming together

* Change accessibility of Infinity

Might seem more convoluted, but now Infinity doesn't carry its own instance as static.

* Change the keyboard mixin, and key event

Having packages for the events would be nice.

* Add OverlayEvent and rendering of modules

I don't like Lambdas needing final stuff, but I get synchronization

* Add tick event, update mixin, and fix Ghost not registering mods

Ghost is always initially false, and the config doesn't load until after LoadEvent is invoked, meaning Ghost misses it every time.

* Change instance access

What was I thinking..

* Add Panic and Sprint

Panic is fairly straightforward, disables all active mods. Sprint internally invokes the toggled sprint accessibility setting, which inherently does the exact thing sprint wants.

* Update Efkolia

* Create Keyboard GLFW utility, update Sprint impl

Sprint is not accurate if depending on a toggle, and can fail if the sprint key is unbound. This fixes it. Furthermore, it seems not all  GLFW keys can be printed out through GLFW.getKeyName(), as such I'm forced to resort to an enum type thing.

* Rename utils package

* Add SafeWalk and Fullbright

I'm noticing that I haven't used events. I feel that mixins are significantly more effective, and will try to not use events. Of course, for areas where multiple things must occur, I will consider events, however I'm thinking we could deprecate the event system in its entirety

* Remove LoadEvent

Ghost will deal with that once its used

* Added NoFOV, update mixin docs

* Prevent module binds from toggling when typing in chat

We might just make this a KeyEvent level check instead

* Add a keyevent todo

* Prevent binds from firing in textboxes

This will forbid binds from firing if a screen is active. This is any possible ui display, like the inventory, chat, or options menu

* Resolve todo

* Force gamma lighting higher

This makes it very bright, which is kinda the point

* Implement a GUI for modules

This will continuously be populated with the newer mods as we continue development

* Prevent Ghost from being visible in GUI

It makes more sense for it to be a bind only thing

* Implement AutoClicker

* Finalize some modules

Style guide thing

* Implement a simple NameProtect

We will make it into the complex obfuscation thing, but later on

* Implement a Command system + GUI

This works significantly better than the previous system, so that's nice

* Update fabric dependencies

* Prevent console from displaying outside of game

Locations such as the main menu are a no go.

* Update gradle dependencies

* Add an uptime in the loader, update commands

* Add command usages

* Fix help command

Logger is deprecated thanks to the new Loader changes... nice

* Refactor packages, introduce deprecated InfinityScreen

Hmm, thinking about whether I even need this

* Update accessibility

Not happy about this..

* Introduce TextRenderer for managed rendering

This isn't for text alone, but rather for text on certain sides of the screen

* Add executor registration

Forgot about this

* Update Efkolia

* Fix annotation positions

* Update README, add PacketEvent, update doAttack in mixins

I'm annoyed that I missed this for AutoClicker

* Bugfix AutoClicker

Resetting the cooldown all the time is not what we want, which is what we got thanks to displace in mixins

* Improve command clarity

This gives better feedback

* Fabric update (#12)

* update to 1.18.2

code is broken in many places, specifically mixins from what i can see.

* migrateMappings from loom

didn't work perfectly but it cleans up a lot of what i need to do

* resolve rendering errors, finalize update

java executable can be deprecated since getResourcesAsStream works as intended now. In fact, java executable itself is broken lol. rendering issuse were being caused by some strange mixin injection not working. i dont know the cause but ive moved the injector very slightly ahead, so it should work exactly as it always has. everything should be working fine

* bump infinity version

cuz why not :)

* resolve strange class ambiguity, update github action

* bring back java executable WILL NOT BUILD

currently efkolia is way out of date. i need to finalize maven publishing over there, and then this code will build correctly in dev. otherwise, the mod itself should work in normal environments.

* update efkolia, fix java executable

* replace deprecated method

funny because internally, StringUtils used this exact call. I don't know lol

* update java ver in mixin

code does not run because of an oversight in Efkolia java executable logic. mod should still work fine outside of development

* update to 1.19, deprecate nameprotect

this is it, im merging this down and forcing a re-write. im fed up with the style of this code, the conventions, and the ridiculousness of minecraft and my codes inflexibility to adapt to it (nameprotect is deprecated for this reason)