implement a runtime tweak loader, and config for tweaks
pretty big change, let's go into detail.

First thing, for the dynamic loader, I opted to make use of Java.NIO FileSystems because (1) newer and modern API and (2) I've used it before in Efkolia and have a somewhat understanding. 

Something that was very beneficial was the fact that we did NOT need to access the files, we just needed to find them and see their names (or paths technically, names were extracted from that). This allowed me to strip a huge chunk out of JavaExecutable from Efkolia and narrow down the specific methods that are necessary to traverse directories. It was important to ensure a jar check since (1) FileSystems fails when trying to pull a newFileSystem from the non-jar source, and (2) a non-jar should theoretically only be possible in a development environment, meaning the class data is in the gradle build folder, and can instead be treated just as a normal defaultFileSystem. I heavily encourage users to read what exactly FileSystem does and why it works the way it does for Zip/Jar files, and it would make more sense why we do what we do. From then on all we do is get the paths of all modules in the impl directory. Rather than hard-coding it I opted to use an already existing module to represent its package, which in this case was AutoClicker. After that its a matter of some string manipulation to create the name required by Class.forName, and we have successfully instantiated. Note that this approach only works in this very specific case. It's something that can only work for a jar that's inspecting itself, not for any other classes outside the jar. 

The config is fairly trivial, by default I have set it so that instances are not created for the tweaks. Reason being Gson creates instances using its fromJson method, so I want to value those instances instead of my own. There is a safety catch within load that will populate null entries with freshly created instances. Afterwards, once loading is guaranteed finished, and the tweaks map is guaranteed populated, a shutdownHook is added to cause the tweak manager to save everything.