allow modded liquid support
my logic for liquid rendering on xray was that only lava should be visible. the reason for this is because lava is a dangerous liquid, whereas water is largely just a mere inconvenience, as a result i prioritised having lava rendered. that being said, this lead to no mod support since any other potential dangerous liquids would also remain hidden, which could lead to a future problem. to rememdy this, i conidered that vanilla mc only has two liquids, water and lava, and changed the check so that if it tried to render water, it would not go through. this lets lava render as normal, and allows for any other liquids as well.