partial cleanup, update loom, find xray incompat (no solution)
It was a painfully annoying time to try and find the issue, but in the end I found it. The BlockRenderDispatcher call in RebuildTask is redirected by the fabric-api, more specifically the indigo renderer they provide. This causes issues since it means the usual BlockRendererDispatcher.renderBatched is never called, so our original method would silently fail. This new mixin will move the redirection directly to the RebuildTask, causing a hard crash due to mixin incompatibility. I WILL FIX THIS, I just need to explore the possible ways that I can go about it.