[REWORK ASAP] implement freecam
ill admit, ive been frustrated over trying to implement this for days, which is why im settling for such a poorly done version of the code. This is really poorly written and breaks almost all the standards I tried to set across this project. I'm placing this in purely because I just want it to be over with, but mark my words, i WILL fix this.